requirements.txt
streamlit==1.28.0
tensorflow==2.16.1
numpy
pillow
requests
scipy